Modified examples and added new rcwa demo script
Introduce new example - achromatic rcwa lens

+++ dflat/physical_optical_layer/core/colburn_tensor_utils.py
@@ -76,12 +76,12 @@ def eig_general(A, eps=1e-6):
         batchSize, pixelsX, pixelsY, Nlay, dim, _ = A.shape
 
         # Calculate intermediate matrices.
-        I = tf.eye(num_rows=dim, dtype=tf.complex128)
+        I = tf.eye(num_rows=dim, dtype=tf.complex64)
         D = tf.reshape(D, shape=(batchSize, pixelsX, pixelsY, Nlay, dim, 1))
         shape_di = (batchSize, pixelsX, pixelsY, Nlay, dim, 1)
         shape_dj = (batchSize, pixelsX, pixelsY, Nlay, 1, dim)
-        E = tf.ones(shape=shape_di, dtype=tf.complex128) * tf.linalg.adjoint(D)
-        E = E - D * tf.ones(shape=shape_dj, dtype=tf.complex128)
+        E = tf.ones(shape=shape_di, dtype=tf.complex64) * tf.linalg.adjoint(D)
+        E = E - D * tf.ones(shape=shape_dj, dtype=tf.complex64)
         E = tf.linalg.adjoint(D) - D
 
         # Lorentzian broadening.
@@ -95,4 +95,5 @@ def eig_general(A, eps=1e-6):
         grad_A = tf.linalg.matmul(tf.linalg.inv(tf.linalg.adjoint(U)), grad_A)
         return grad_A
 
+    # return [eigenvalues, eigenvectors], grad
     return [eigenvalues, eigenvectors], grad
